// src/pages/CustomerManagement/customerDetail/dxl/data.d.ts
export interface RequirementUser {
  my: RequirementData[];
  other: RequirementData[];
}
// ...
export interface RequirementBean {
  id: string;
  req_num: string;
  category: string;
  budget: string;
  merchant: string;
  photo_style: string;
  photo_time: string;
  wedding_style: string;
  hotel: string;
  hotel_star: string;
  hotel_tables: string;
  hotel_hall: string;
  car_brand: string;
  car_series: string;
  car_num: string;
  car_time: string;
  est_arrival_time: string;
  status: string;
  status_txt: string;
  create_time: string;
  update_time: string;
  user_id: string;
  customer_id: string;
  user_name: string;
  remark: string;
  city_info: CityInfo;
  city_code: string;
  category_txt: string;
  phase: string;
}

export interface CityInfo {
  province; string;
  city: string;
  district: string;
  full: string;
  city_code: string;
}

export interface RequirementData {
  category_id: string;
  category_name: string;
  data: RequirementBean[];
}
// ...
export interface ProjectTeam {
  req_id: string;
  team_name: string;
  category: string;
  category_text: string;
  owner: string;
  owner_id: string;
  team_member: TeamMember[];
}

export interface TeamMember {
  id: string;
  team_user_id: string;
  team_key_name: string;
  team_user_name: string;
}
// ...
export interface UserEntity {
  name: string;
  id: string;
  job_number: string;
  company_name: string;
  structure_name: string;
}